fix advanced mode styling
