+++ packages/editor/src/components/styles.module.scss
@@ -97,14 +97,22 @@
   }
 }
 
-.hierarchyTab {
+.dockableTab {
   display: flex;
 
-  .hierarchyTabButtons {
+  .dockableTabButtons {
     display: flex;
     position: absolute;
     width: 100%;
     padding-left: 110px;
     align-items: center;
+
+    button {
+      margin: auto;
+      background: none;
+      border: none;
+      display: flex;
+      align-items: center;
+    }
   }
 }
